requirements.txt
alembic==1.4.3
Babel==2.9.0
Flask==2.1.0
gunicorn==20.0.4
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
requests==2.22.0
psycopg2
psycopg2-binary==2.8.2
Flask-Mail==0.9.1
python-dotenv
Flask-SQLAlchemy==2.1
SQLAlchemy==1.3.12
tensorflow==2.6.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Jinja2==3.0
matplotlib==3.5.2
mysql-connector-python==8.0.26
mysqlclient==2.0.3
itsdangerous==2.0
six==1.15.0
Werkzeug==2.0